+++ bootcfg/http/metadata.go
@@ -2,6 +2,7 @@ package http
 
 import (
 	"fmt"
+	"io"
 	"net/http"
 	"strings"
 
@@ -39,9 +40,32 @@ func (s *Server) metadataHandler() ContextHandler {
 		}
 
 		w.Header().Set(contentType, plainContentType)
-		for key, value := range data {
-			fmt.Fprintf(w, "%s=%v\n", strings.ToUpper(key), value)
-		}
+		renderAsEnvFile(w, "", data)
 	}
 	return ContextHandlerFunc(fn)
+}
+
+// renderAsEnvFile writes map data into a KEY=value\n "env file" format,
+// descending recursively into nested maps and prepending parent keys.
+//
+// For example, {"outer":{"inner":"val"}} -> OUTER_INNER=val). Note that
+// structure is lost in this transformation, the inverse transfom has two
+// possible outputs.
+func renderAsEnvFile(w io.Writer, prefix string, root map[string]interface{}) {
+	for key, value := range root {
+		name := prefix + key
+		switch val := value.(type) {
+		case string, bool, float64:
+			// simple JSON unmarshal types
+			fmt.Fprintf(w, "%s=%v\n", strings.ToUpper(name), val)
+		case map[string]string:
+			m := map[string]interface{}{}
+			for k, v := range val {
+				m[k] = v
+			}
+			renderAsEnvFile(w, name+"_", m)
+		case map[string]interface{}:
+			renderAsEnvFile(w, name+"_", val)
+		}
+	}
 }